requirements.txt
Flask>=2
gunicorn==20.1.0
tensorflow-cpu>=2.0.0